start restructuring
