+++ Gruntfile.js
@@ -124,16 +124,14 @@ module.exports = function(grunt) {
           'src/DistortableImageOverlay.js',
           'src/DistortableCollection.js',
           'src/edit/getEXIFdata.js',
-          'src/edit/EditHandle.js',
-          'src/edit/LockHandle.js',
-          'src/edit/DistortHandle.js',
-          'src/edit/RotateScaleHandle.js',
-          'src/edit/RotateHandle.js',
-          'src/edit/ScaleHandle.js',
+          'src/edit/handles/EditHandle.js',
+          'src/edit/handles/*.js',
           'src/iconsets/IconSet.js',
           'src/iconsets/KeymapperIconSet.js',
           'src/iconsets/ToolbarIconSet.js',
           'src/edit/tools/EditAction.js',
+          'src/edit/tools/EditAction.*',
+          'src/edit/tools/*Action.js',
           'src/edit/tools/DistortableImage.PopupBar.js',
           'src/edit/tools/DistortableImage.ControlBar.js',
           'src/edit/DistortableImage.Edit.js',
