+++ examples/select.html
@@ -82,7 +82,7 @@
 
         img4 = L.distortableImageOverlay(
           'example.jpg', {
-            actions: [ToggleRotate, ToggleScale, ToggleRotateScale, ToggleLock, ToggleTransparency, Delete],
+            actions: [L.RotateAction, L.ScaleAction, L.EditAction.ToggleRotateScale, L.EditAction.ToggleLock, L.OpacityAction, L.DeleteAction],
             corners: [
               L.latLng(51.51, -0.07),
               L.latLng(51.51,-0.03),
