+++ src/edit/handles/EditHandle.js
