+++ src/edit/handles/LockHandle.js
