+++ src/edit/handles/RotateHandle.js
