+++ src/edit/handles/RotateScaleHandle.js
