+++ src/edit/handles/ScaleHandle.js
