+++ src/edit/tools/DistortableImage.PopupBar.js
@@ -3,305 +3,6 @@ L.distortableImage = L.DistortableImage;
 
 L.DistortableImage.action_map = {};
 
-var ToggleTransparency = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var edit = overlay.editing,
-        use, tooltip;
-
-    if (edit._transparent) {
-      use = 'opacity';
-      tooltip = 'Make Image Opaque';
-    } else {
-      use = 'opacity_empty';
-      tooltip = 'Make Image Transparent';
-    }
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: tooltip,
-    };
-
-    L.DistortableImage.action_map.t = '_toggleTransparency';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    L.IconUtil.toggleXlink(this._link, 'opacity_empty', 'opacity');
-    L.IconUtil.toggleTooltip(this._link,
-        'Make Image Opaque',
-        'Make Image Transparent');
-    editing._toggleTransparency();
-  },
-});
-
-var ToggleOutline = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var edit = overlay.editing,
-        use, tooltip;
-
-    if (edit._outlined) {
-      use = 'border_clear';
-      tooltip = 'Remove Border';
-    } else {
-      use = 'border_outer';
-      tooltip = 'Add Border';
-    }
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: tooltip,
-    };
-
-    L.DistortableImage.action_map.o = '_toggleOutline';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    L.IconUtil.toggleXlink(this._link, 'border_clear', 'border_outer');
-    L.IconUtil.toggleTooltip(this._link, 'Remove Border', 'Add Border');
-    editing._toggleOutline();
-  },
-});
-
-var Delete = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var use = 'delete_forever';
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: 'Delete Image',
-    };
-
-    L.DistortableImage.action_map.Backspace = '_removeOverlay'; // backspace windows / delete mac
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    editing._removeOverlay();
-  },
-});
-
-var ToggleLock = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var edit = overlay.editing,
-        use, tooltip;
-
-    if (edit._mode === 'lock') {
-      use = 'unlock';
-      tooltip = 'Unlock';
-    } else {
-      use = 'lock';
-      tooltip = 'Lock';
-    }
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: tooltip,
-    };
-
-    L.DistortableImage.action_map.l = '_toggleLock';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    L.IconUtil.toggleXlink(this._link, 'unlock', 'lock');
-    L.IconUtil.toggleTooltip(this._link, 'Unlock', 'Lock');
-    editing._toggleLock();
-  },
-});
-
-var ToggleRotateScale = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var edit = overlay.editing,
-        use, tooltip;
-
-    if (edit._mode === 'rotateScale') {
-      use = 'distort';
-      tooltip = 'Distort Image';
-    } else {
-      use = 'crop_rotate';
-      tooltip = 'Rotate+Scale Image';
-    }
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: tooltip,
-    };
-
-    L.DistortableImage.action_map.d = '_toggleRotateScale';
-    L.DistortableImage.action_map.r = '_toggleRotateScale';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    L.IconUtil.toggleXlink(this._link, 'distort', 'crop_rotate');
-    L.IconUtil.toggleTooltip(this._link, 'Distort Image', 'Rotate+Scale Image');
-    editing._toggleRotateScale();
-  },
-});
-
-var Export = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var use = 'get_app';
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: 'Export Image',
-    };
-
-    L.DistortableImage.action_map.e = '_getExport';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    editing._getExport();
-  },
-});
-
-var ToggleOrder = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var edit = overlay.editing,
-        use, tooltip;
-
-    if (edit._toggledImage) {
-      use = 'flip_to_front';
-      tooltip = 'Stack to Front';
-    } else {
-      use = 'flip_to_back';
-      tooltip = 'Stack to Back';
-    }
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: tooltip,
-    };
-
-    L.DistortableImage.action_map.j = '_toggleOrder';
-    L.DistortableImage.action_map.k = '_toggleOrder';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    L.IconUtil.toggleXlink(this._link, 'flip_to_front', 'flip_to_back');
-    L.IconUtil.toggleTooltip(this._link, 'Stack to Front', 'Stack to Back');
-    editing._toggleOrder();
-  },
-});
-
-var EnableEXIF = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var use = 'explore';
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: 'Geolocate Image',
-    };
-
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var image = this._overlay.getElement();
-
-    // eslint-disable-next-line new-cap
-    EXIF.getData(image, L.EXIF(image));
-  },
-});
-
-var Revert = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var use = 'restore';
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: 'Restore Original Image Dimensions',
-    };
-
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    this._overlay._revert();
-  }
-});
-
-var ToggleRotate = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var use = 'rotate';
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: 'Rotate Image',
-    };
-
-    L.DistortableImage.action_map.CapsLock = '_toggleRotate';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    editing._toggleRotate();
-  },
-});
-
-var ToggleScale = L.EditAction.extend({
-  initialize: function(map, overlay, options) {
-    var use = 'scale';
-
-    options = options || {};
-    options.toolbarIcon = {
-      svg: true,
-      html: use,
-      tooltip: 'Scale Image',
-      className: 'ToggleScale'
-    };
-    
-    L.DistortableImage.action_map.s = '_toggleScale';
-    L.EditAction.prototype.initialize.call(this, map, overlay, options);
-  },
-
-  addHooks: function() {
-    var editing = this._overlay.editing;
-
-    editing._toggleScale();
-  },
-});
-
 L.DistortableImage.PopupBar = L.Toolbar2.Popup.extend({
   options: {
     anchor: [0, -10],
@@ -321,60 +22,60 @@ L.DistortableImage.PopupBar = L.Toolbar2.Popup.extend({
     // ]
   },
 
-  hasScale: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(ToggleScale) === 1;
-  },
+  // hasScale: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(ToggleScale) === 1;
+  // },
 
-  hasRotate: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(ToggleRotate) === 1;
-  },
+  // hasRotate: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(ToggleRotate) === 1;
+  // },
 
-  hasRotateScale: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(ToggleRotateScale) === 1;
-  },
+  // hasRotateScale: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(ToggleRotateScale) === 1;
+  // },
 
-  hasTransparency: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(ToggleTransparency) === 1;
-  },
+  // hasTransparency: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(ToggleTransparency) === 1;
+  // },
 
-  hasRevert: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(Revert) === 1;
-  },
+  // hasRevert: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(Revert) === 1;
+  // },
 
-  hasDelete: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(Delete) === 1;
-  },
+  // hasDelete: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(Delete) === 1;
+  // },
 
-  hasOrder: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(ToggleOrder) === 1;
-  },
+  // hasOrder: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(ToggleOrder) === 1;
+  // },
 
-  hasLock: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(ToggleLock) === 1;
-  },
+  // hasLock: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(ToggleLock) === 1;
+  // },
 
-  hasEXIF: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(EnableEXIF) === 1;
-  },
+  // hasEXIF: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(EnableEXIF) === 1;
+  // },
 
-  hasExport: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(Export) === 1;
-  },
+  // hasExport: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(Export) === 1;
+  // },
 
-  hasOutline: function() {
-    var overlay = this._arguments[1];
-    return overlay.editing.editActions.indexOf(ToggleOutline) === 1;
-  }
+  // hasOutline: function() {
+  //   var overlay = this._arguments[1];
+  //   return overlay.editing.editActions.indexOf(ToggleOutline) === 1;
+  // }
 });
 
 L.distortableImage.popupBar = function(latlng, options) {
@@ -384,14 +85,14 @@ L.distortableImage.popupBar = function(latlng, options) {
 L.DistortableImageOverlay.addInitHook(function() {
   /** Default actions */
   this.ACTIONS = [
-    ToggleTransparency,
-    ToggleOutline,
-    ToggleLock,
-    ToggleRotateScale,
-    ToggleOrder,
-    Revert,
-    Export,
-    Delete,
+    L.OpacityAction,
+    L.EditAction.ToggleOutline,
+    L.ToggleLockAction,
+    L.EditAction.ToggleRotateScale,
+    L.EditAction.ToggleOrder,
+    L.RevertAction,
+    L.ExportAction,
+    L.DeleteAction,
   ];
 
 if (this.options.actions) { /* (`this` being DistortableImageOverlay, not the toolbar) */
